Push to work laptop [Sqs-consumer] [Working]

+++ src/main.js
@@ -5,19 +5,56 @@ const awsS3Utility = require("/Users/ramkishoremadheshwaran/WebstormProjects/rmk
 const mongoDataHandler = require ("/Users/ramkishoremadheshwaran/WebstormProjects/rmkshr_javascript/src/mongo-data-handler.js");
 
 
-awsS3Utility.getAllValidFileNames.then(nameList => {
-    nameList.forEach(fileName => {
-            awsS3Utility.readFileDataFromS3(fileName).then(jsonData =>{
+// awsS3Utility.getAllValidFileNames.then(nameList => {
+//     nameList.forEach(fileName => {
+//             awsS3Utility.readFileDataFromS3(fileName).then(jsonData =>{
+//                 mongoDataHandler.insertMany(jsonData, fileName).then(fileName => {
+//                     // There is no AWS provided API to rename a file in S3 bucket, so we make a copy and give our new name,
+//                     // upload that file to the same bucket and delete the older file.
+//                     awsS3Utility.copyS3BucketFile(fileName).then(fileName => {
+//                         awsS3Utility.deleteS3BucketFile(fileName);
+//                     });
+//                 });
+//             });
+//         });
+// }).catch(exception => console.log(exception));
+
+
+const { Consumer } = require('sqs-consumer');
+
+const app = Consumer.create({
+    queueUrl: 'https://sqs.us-east-1.amazonaws.com/418090287981/s3tosqsstacktest-StandardQueue-2T8hyvzfPfMZ',
+    handleMessage: async (message) => {
+        console.log(message);
+        const data = message["Body"];
+        const jsonData = JSON.parse(data);
+        const records = jsonData.Records;
+        const recordData = records[0];
+        const s3data = recordData.s3;
+        const fileObject = s3data.object;
+        const fileName = fileObject.key;
+        console.log(fileName);
+        if(fileName.startsWith("New")) {
+            await awsS3Utility.readFileDataFromS3(fileName).then(jsonData => {
                 mongoDataHandler.insertMany(jsonData, fileName).then(fileName => {
-                    // There is no AWS provided API to rename a file in S3 bucket, so we make a copy and give our new name,
-                    // upload that file to the same bucket and delete the older file.
                     awsS3Utility.copyS3BucketFile(fileName).then(fileName => {
                         awsS3Utility.deleteS3BucketFile(fileName);
                     });
                 });
-            });
-        });
-}).catch(exception => console.log(exception));
+            }).catch(exception => console.log(exception));
+        }
+    }
+});
+
+app.on('error', (err) => {
+    console.error(err.message);
+});
+
+app.on('processing_error', (err) => {
+    console.error(err.message);
+});
+
+app.start();
